complete quantized op family split
