+++ onnx2tf/tflite_builder/op_builders/qlinear_concat.py
@@ -5,7 +5,7 @@ from typing import Any
 import numpy as np
 
 from onnx2tf.tflite_builder.ir import OperatorIR
-from onnx2tf.tflite_builder.op_builders.quantized import (
+from onnx2tf.tflite_builder.op_builders.quantized_common import (
     _normalize_axis,
     _promote_internal_uint8_tensor_to_int8,
     _require_const,
