+++ onnx2tf/tflite_builder/op_builders/qlinear_fc.py
@@ -5,7 +5,7 @@ from typing import Any
 import numpy as np
 
 from onnx2tf.tflite_builder.ir import OperatorIR, QuantParamIR
-from onnx2tf.tflite_builder.op_builders.quantized import (
+from onnx2tf.tflite_builder.op_builders.quantized_common import (
     _add_scalar_onnx_requantization,
     _normalize_quant_params,
     _promote_internal_uint8_tensor_to_int8,
